requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.0.1
anyio==3.6.2
appnope==0.1.3
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.26.76
botocore==1.29.76
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.25
click==7.1.2
clickhouse-connect==0.5.24
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
cryptography==40.0.2
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
deeplake==3.6.7
dill==0.3.6
dnspython==2.3.0
duckdb==0.8.0
entrypoints==0.4
executing==1.2.0
fastapi==0.95.2
flatbuffers==23.5.9
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.48.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.59.0
h11==0.14.0
hnswlib==0.7.0
httplib2==0.22.0
httptools==0.5.0
humanfriendly==10.0
humbug==0.3.1
icrawler==0.6.6
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.221
langchainplus-sdk==0.0.19
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mutagen==1.46.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.14.1
openai==0.27.7
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathos==0.3.0
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
platformdirs==3.8.0
posthog==3.0.1
pox==0.3.2
ppft==1.7.6.6
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.11.0

pydeck==0.8.1b0
pyfiglet==0.8.post1
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.11.1
PyPDF2==3.0.1
PyPDF4==1.27.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.21
python-resize-image==1.1.20
pytube==15.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.0
regex==2023.5.5
requests==2.31.0
rich==13.4.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
stack-data==0.6.2
starlette==0.27.0
streamlit==1.24.0
sympy==1.12
tenacity==8.2.2
termcolor==1.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
Wand==0.6.11
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
wikipedia==1.4.0
wrapt==1.15.0
XlsxWriter==3.1.2
yarl==1.9.2
youtube-transcript-api==0.6.1
yt-dlp==2023.6.22
zipp==3.15.0
zstandard==0.21.0
